One water-quality vocabulary for the WHO screen and the indices
The live run at the Potomac said "none of the sampled parameters has a WHO
guideline value" in the WHO screen and then indexed dissolved oxygen in the
same report: the screen matched raw labels against its own keys, so USGS's
"DO" was invisible to it. The names, aliases, canonical units and conversions
move to aquascope.utils.parameters (no pandas, light enough for the workbench
to import), the indices import them from there, and who_screen resolves every
label through the same vocabulary and converts values to the guideline's unit
before comparing (arsenic in ug/L against 0.01 mg/L).
